Added version of string formatter using scan; split formatters off into seperate modules; added benchmarking tests
