+++ bench/regex_formatter_bench.exs
@@ -1,4 +1,20 @@
 defmodule RegexFormatterBench do
-  use BenchFella
+  use Benchfella
 
+  setup_all do
+    depth = :erlang.system_flag(:backtrace_depth, 100)
+    {:ok, depth}
+  end
+
+  teardown_all depth do
+    :erlang.system_flag(:backtrace_depth, depth)
+  end
+
+  bench "Original format function recursing through string" do
+    NaturalSort.FormatViaRecursion.format_item("foo123foo456_#05", false)
+  end
+
+  bench "New format function using scan" do
+    NaturalSort.FormatViaScan.format_item("foo123foo456_#05", false)
+  end
 end
