+++ bench/snapshots/2015-05-15T16:25:44.snapshot
@@ -1,0 +1,3 @@
+duration:1.0;mem stats:false;sys mem stats:false
+module;test;tags;iterations;elapsed
+RegexFormatterBench	Original format function recursing through string		50000	2005240
