+++ bench/snapshots/2015-05-15T16:25:53.snapshot
@@ -1,0 +1,2 @@
+duration:1.0;mem stats:false;sys mem stats:false
+module;test;tags;iterations;elapsed
