+++ bench/snapshots/2015-05-15T17:17:09.snapshot
@@ -1,0 +1,4 @@
+duration:1.0;mem stats:false;sys mem stats:false
+module;test;tags;iterations;elapsed
+RegexFormatterBench	Original format function recursing through string		50000	1963134
+RegexFormatterBench	New format function using scan		50000	2129754
