+++ bench/snapshots/2015-05-15T17:57:54.snapshot
@@ -1,0 +1,4 @@
+duration:1.0;mem stats:false;sys mem stats:false
+module;test;tags;iterations;elapsed
+RegexFormatterBench	Original format function recursing through string		50000	1916309
+RegexFormatterBench	New format function using scan		50000	2115403
